added rooms folder under API
added rooms folder under the API folder to handle room schema api requests
